refactor(abg): clean up structure
