--- action-binding-generator/src/main/kotlin/io/github/typesafegithub/workflows/actionbindinggenerator/TextTransformation.kt
@@ -1,7 +1,0 @@
-package io.github.typesafegithub.workflows.actionbindinggenerator
-
-internal fun String.toKotlinPackageName(): String =
-    replace("-", "")
-        .lowercase()
-
-internal fun String.removeTrailingWhitespacesForEachLine() = lines().joinToString(separator = "\n") { it.trimEnd() }
