+++ automation/code-generator/src/main/kotlin/io/github/typesafegithub/workflows/codegenerator/ActionsMetadataReading.kt
@@ -1,7 +1,7 @@
 package io.github.typesafegithub.workflows.codegenerator
 
-import io.github.typesafegithub.workflows.actionbindinggenerator.ActionCoords
-import io.github.typesafegithub.workflows.actionbindinggenerator.prettyPrint
+import io.github.typesafegithub.workflows.actionbindinggenerator.domain.ActionCoords
+import io.github.typesafegithub.workflows.actionbindinggenerator.domain.prettyPrint
 import io.github.typesafegithub.workflows.codegenerator.model.ActionBindingRequest
 import java.nio.file.Files
 import java.nio.file.Path
