+++ automation/code-generator/src/main/kotlin/io/github/typesafegithub/workflows/codegenerator/model/ActionBindingRequest.kt
@@ -1,6 +1,6 @@
 package io.github.typesafegithub.workflows.codegenerator.model
 
-import io.github.typesafegithub.workflows.actionbindinggenerator.ActionCoords
+import io.github.typesafegithub.workflows.actionbindinggenerator.domain.ActionCoords
 
 data class ActionBindingRequest(
     val actionCoords: ActionCoords,
