+++ automation/code-generator/src/test/kotlin/io/github/typesafegithub/workflows/codegenerator/AddDeprecationInfoTest.kt
@@ -1,6 +1,6 @@
 package io.github.typesafegithub.workflows.codegenerator
 
-import io.github.typesafegithub.workflows.actionbindinggenerator.ActionCoords
+import io.github.typesafegithub.workflows.actionbindinggenerator.domain.ActionCoords
 import io.github.typesafegithub.workflows.codegenerator.model.ActionBindingRequest
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
